requirements.txt
Django==5.0.2
djangorestframework
djangorestframework-simplejwt
sqlparse
psycopg2-binary